requirements.txt
python-telegram-bot==20.7
openai==1.12.0
python-dotenv==1.0.0
gTTS==2.4.0
requests==2.31.0
pydub==0.25.1
elevenlabs==2.0.0 